Refactoring DbAccess class
