+++ client/client/MainWindow.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows;
+using System.Collections.Generic;
 using System.Windows.Controls;
-using System.Windows.Input;
-using System.ComponentModel;
+using System.Data;
 
 namespace client
 {
@@ -10,11 +10,26 @@ namespace client
     /// </summary>
     public partial class MainWindow : Window
     {
+        DbAccess db;
+        Dictionary<string, string> queries = new Dictionary<string, string>()
+        {
+            {"Employee", "SELECT Employee.Id As Id, Employee.Name As Name, Department.Name as Department, Type " +
+                "FROM Employee LEFT JOIN Department ON Department.Id = Employee.Department ORDER BY Employee.Id;"},
+
+            {"Default", "SELECT * FROM "},
+
+            {"Position", "SELECT * FROM [Position] ORDER BY Id" },
+
+            {"Infrastructure", "SELECT Infrastructure.Id AS Id, Inventory.Name AS Name, Infrastructure.DateRelease AS Released, Infrastructure.DatePurchase AS Purchased, Office.Name AS Office, Employee.Name AS Responsible, Infrastructure.Price AS Price FROM Employee RIGHT JOIN (Office RIGHT JOIN (Inventory RIGHT JOIN Infrastructure ON Inventory.[Id] = Infrastructure.[Name]) ON Office.[Id] = Infrastructure.[Office]) ON Employee.Id = Infrastructure.[Responsible] ORDER BY Infrastructure.Id;"},
+
+            {"Office","SELECT Office.Id, Department.Name AS Department, Office.Name AS Name FROM Office LEFT JOIN Department ON Department.Id = Office.Department ORDER BY Office.Id;" }
+        };
         public MainWindow()
         {
             InitializeComponent();
-            DbAccess bd = new DbAccess() {dg = dataGrid};
-            bd.SetTree(treeView);
+            db = new DbAccess() { conString = "Provider=Microsoft.ACE.OLEDB.16.0;Data Source=DataBase.accdb;" };
+            TreeView treeView = db.SetTree("Tables");
+            treeStack.Children.Add(treeView);
 
         }
 
@@ -22,6 +37,15 @@ namespace client
         {
             EditMode em = new EditMode();
             em.Show();
+        }
+        private void Item_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
+        {
+            TreeViewItem item = (TreeViewItem)sender;
+            string content = item.Header.ToString();
+            if (queries.ContainsKey(content))
+                dataGrid.ItemsSource = db.MakeQuery(queries[content]);
+            else
+                dataGrid.ItemsSource = db.MakeQuery(queries["Default"] + content);
         }
     }
 }
