+++ client/client/classes/DbAccess.cs
@@ -7,22 +7,9 @@ namespace client
 {
     internal class DbAccess
     {
-        static string conString = "Provider=Microsoft.ACE.OLEDB.16.0;Data Source=DataBase.accdb;";
-        public DataGrid dg;
-        Dictionary<string, string> queries = new Dictionary<string, string>()
-        {
-            {"Employee", "SELECT Employee.Id As Id, Employee.Name As Name, Department.Name as Department, Type " +
-                "FROM Employee LEFT JOIN Department ON Department.Id = Employee.Department ORDER BY Employee.Id;"},
-
-            {"Default", "SELECT * FROM "},
-
-            {"Position", "SELECT * FROM [Position] ORDER BY Id" },
-
-            {"Infrastructure", "SELECT Infrastructure.Id AS Id, Inventory.Name AS Name, Infrastructure.DateRelease AS Released, Infrastructure.DatePurchase AS Purchased, Office.Name AS Office, Employee.Name AS Responsible, Infrastructure.Price AS Price FROM Employee RIGHT JOIN (Office RIGHT JOIN (Inventory RIGHT JOIN Infrastructure ON Inventory.[Id] = Infrastructure.[Name]) ON Office.[Id] = Infrastructure.[Office]) ON Employee.Id = Infrastructure.[Responsible] ORDER BY Infrastructure.Id;"},
-
-            {"Office","SELECT Office.Id, Department.Name AS Department, Office.Name AS Name FROM Office LEFT JOIN Department ON Department.Id = Office.Department ORDER BY Office.Id;" }
-        };
-        public static DataTable MakeQuery(string query)
+        public string conString;
+        
+        public DataView MakeQuery(string query)
         {
             OleDbConnection con = new OleDbConnection(conString);
 
@@ -33,44 +20,33 @@ namespace client
             DataTable dt = new DataTable();
             oda.Fill(dt);
             con.Close();
-
-            return dt;
-//            dg.ItemsSource = dt.DefaultView;
+            DataView dview = new DataView(dt);
+            return dview;
             
         }
-        public void SetTree(TreeView tree)
+        public TreeView SetTree(string branch)
         {
+            TreeView tree = new TreeView();
             OleDbConnection con = new OleDbConnection(conString);
-
-            con.Open();
-
             string[] restrictions = new string[4];
             restrictions[3] = "Table";
 
-            DataTable dt = con.GetSchema("Tables", restrictions);
+            con.Open();
+            DataTable dt = con.GetSchema(branch , restrictions);
             con.Close();
-
-            TreeViewItem treeTable = new TreeViewItem() { Header = "Таблица" };
+            
+            TreeViewItem treeItem = new TreeViewItem() { Header = branch };
 
             foreach (DataRow row in dt.Rows)
             {
                 TreeViewItem item = new TreeViewItem();
                 item.Header = row["TABLE_NAME"].ToString();
-                item.MouseDoubleClick += Item_MouseDoubleClick;
-                treeTable.Items.Add(item);
+                //item.MouseDoubleClick += Item_MouseDoubleClick;
+                treeItem.Items.Add(item);
             }
-            tree.Items.Add(treeTable);
-        }
+            tree.Items.Add(treeItem);
 
-        private void Item_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
-        {
-            TreeViewItem item = (TreeViewItem)sender;
-            string content = item.Header.ToString();
-            if (queries.ContainsKey(content))
-                dg.ItemsSource = MakeQuery(queries[content]).DefaultView;
-            else
-                dg.ItemsSource = MakeQuery(queries["Default"] + content).DefaultView;
-
+            return tree;
         }
 
     }
